Refactor engine publishing
Removed the engine publisher to rely on scenes for publishing.
To do this, the engine's callbacks are subscribed each time a
scene is pushed. This required a refactor of the push and replace
functions. Replace now calls pop and push in succession. In
addition, some error handling in push allows you to call it with
a scene instead of an event.