Controller and Publisher Refactor
The controller is built to be generic, requiring a middleware
to access the hardware. Added to events are a series of events
to use for building middleware.

Publishers refactors to use sets instead of dictionaries,
removing the need for an "identity" to track subscribers. The
callback functions operate as the unique identifier.