Particle Class, Queue enhancement, Publisher Fix
Particle class extends Renderable. Will be directly related to
the Emitter class. Intended as a base class to fix the API between
Emitters and Particles.

Queue enhanced with an extend method to reduce the requirement of
for loops to push multiple items to the queue.

Publisher would break when unsubscribing during a publish action.
Added a method to delay unsubscribe to before the next publishing
cycle.