Emitter Class, Publisher Fix
Emitter class has an activate callback that listens for an
event and sets to active state. Also listens for a tick to
activate an emit function.

Publish would break if adding a callback in the middle of a
publishing loop. Added a subscribe_requests queue, and an add
function that is called before publishing. API has not changed.

Removed various logging and unnecessary documentation.