Collision Detection, and Start of Component Refactor
Added a Physics object that applies collision detection and raises
a collision event. Eventually will also include forces which will
be applied to physics objects.

Started a refactor of subscribtion model. Instead of subscribing
future objects will raise an event with a list of their
subscriptions. They raise a related event when they are destroyed.