Extracted View from Renderable and Minor Refactors
View listens for object creation and instantiates Sprites as
needed.

Refactored core engine to push a Tick event when the queue is
empty instead of listening to a Tick to raise a Tick.

Now the engine doesn't need to know that the scene has a subscribe
function just a publish function.