Adds tests with initial design of new API
We don't like the old camera, it's fiddly and hard for users to reason about.
This new API focuses on the actual dimensions of the camera as the frame.
The viewport is defined in percentages of the physical system. This way users
can think in game space and let the engine handle translating back and forth.