engine: Use a simpler representation of statistics during execution
Avoid loading the pandas library and interacting with it until the game ends.

This results in a 10× performance improvement (with profiling turned on),
i.e. a reduction in instrumentation overhead of at least 90%.

Moreover, pandas creates reference cycles which need to be cleaned up by the
garbage collector, falsifying the GC-related measurements.